Migrated over error summary POM and feature to new format
Minor tweaks to structure of load validator
