+++ testing/features/step_definitions/error_summary_steps.rb
@@ -1,18 +1,17 @@
 # frozen_string_literal: true
 
-Given("an Error Summary component is on the page") do
-  @component = DesignSystem::ErrorSummary.new
-  @component.load
+Given("an Example Error Summary component is on the page") do
+  @component = ErrorSummary::Example.new.tap(&:load)
 end
 
-Then("the component will indicate there has been a problem") do
-  expect(@component.initial_form).to have_heading
+Then("the component indicates there is a problem") do
+  expect(@component).to have_heading
 end
 
-Then("the component should list the issues in bullet point format") do
-  expect(@component.initial_form).to have_errors
+Then("the component lists the issues in bullet point format") do
+  expect(@component).to have_errors
 end
 
-Then("the number of errors should equal the number of bullet points") do
-  expect(@component.initial_form.error_quantity).to eq(@component.initial_form.errors.length)
+Then("there is an error for each bullet point") do
+  expect(@component.error_quantity).to eq(@component.errors.length)
 end
